v1.1
Now support toggling switch at scheduled future time based on calendar event start and end times.  If a calendar event match is found and the start time is after the current time, a trigger will be scheduled to toggle the switch at the calendar event start and end times.  Each time refresh and poll are executed, any scheduled are removed and re-evaluated based on the calendar events and search criteria.