v2.2.3
Fixed bug with caching of events
Fixed bug with getting end date of search range and cast error.  If required endTimeHours preference is not set for some reason, it will default to end of current day.
Added toggle to hide offset preferences to help reduce confusion on this optional feature.