v2.3.1
Added an option to use Google Query Matching. By default calendar event matching is done by the HE hub and it allows multiple search strings. Caching of events is not supported when using Google query matching.
Added ability to prefix search words with a "-" (minus) to exclude events with this word in the title or location.
Added ability to exclude All Day Events (default is to include)
Added ability to set whether sequential events will expand the scheduling of when the child switch turns off.  If this is turned off, the switch will turn off 1 minute in advance of calendar event end time to ensure the toggle can happen.  You can override this via the Event End Offset setting.
Added ability to toggle/sync additional switches. A list of switches can be designated to synchronize and a list of switches can be designated to reverse the switch state as the child calendar switch is toggled.