v2.5.1.1
New Feature: If Event End Offset is entered, this amount of time will be included when querying for new calendar events.  For example if you entered an offset of 30 minutes so the switch stays on 30 additional minutes, during a calendar search an additional -30 minutes is added to the query so that any events that just ended 30 minutes ago are still included so the switch will not turn off. This is useful for frequent periodic queries.

Bug Fixes: The "Expand end date for sequential events" included in v2.X only worked for HE based event matching.  Corrected the code to also work when Google Query Matching is enabled.

Got error testing HPM, fixing manifest.