v4.6.0
Added ability to turn on and off other existing switches in the Additional Actions settings. A child switch is no longer required to do so. If this was setup previously, the upgrade process will move the settings to Additional Actions.

Added the ability to enter '\n' within a Gmail notification message which will cause a carriage return within the email sent. In recent versions HTML is automatically stripped so this is a work around for new lines within email notifications.

When selecting switches in the various Additional Actions settings, the filter box is now available to quickly find specific switches.