v4.6.4
Enhanced the Additional Actions \ Text to Parse feature. Previously it would look for text that starts with text entered in the Text Prefix column and set the designated variable to the remaining text. Now it will look for text across the entire line of text and remove the value entered in the Text to Find column and set the designated variable to the remaining text.